feat: added a bids folder for bids.js, and moved all bid-related code to bids.js
